% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parellel_regression.R
\name{parallel_regression}
\alias{parallel_regression}
\title{Parallel computation of multiple regression models}
\usage{
parallel_regression(data, calls)
}
\arguments{
\item{data}{A data frame, or a list of data frames.}

\item{calls}{Character vector, where each element is a function call to a
regression. The \code{data}-argument must always have the value \code{i}.}

\item{cores}{Numeric, the number of cores to use.}
}
\value{
A list of regression model objects.
}
\description{
Parallel computation of multiple regression models.
}
\examples{
library(lme4)
data(sleepstudy)

models = c(
  "lme4::lmer(Reaction ~ 1 + (1 | Subject), data = i)",
  "lme4::lmer(Reaction ~ Days + (Days | Subject), data = i)",
  "lme4::lmer(Reaction ~ Days + (1 | Subject) + (0 + Days | Subject), data = i)"
)

parallel_regression(sleepstudy, models)

}
