% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_longer.R
\name{reshape_longer}
\alias{reshape_longer}
\title{Reshape data into long format}
\usage{
reshape_longer(x, columns = colnames(x), names_to = "key",
  values_to = "value", numeric_timvar = FALSE, id = "id")
}
\arguments{
\item{x}{A data frame.}

\item{columns}{Names of variables (as character vector), or column index of
variables, that should be reshaped. If multiple column groups should be
reshaped, use a list of vectors (see 'Examples').}

\item{names_to}{todo}

\item{values_to}{todo}

\item{numeric_timvar}{todo}

\item{id}{todo}
}
\value{
A reshaped data frame.
}
\description{
\code{reshape_longer()} reshapes one or more columns from
  wide into long format.
}
\examples{
d <- data.frame(
  x = 1:4,
  y1 = rnorm(4),
  y2 = rnorm(4),
  a = c(1, 1, 0, 0),
  b = c(0, 1, 1, 1)
)

# simple reshape of one time-varying variable
reshape_longer(
  d,
  columns = c("y1", "y2"),
  names_to = "time",
  values_to = "new_y"
)

d <- data.frame(
  x = 1:4,
  y1 = rnorm(4),
  y2 = rnorm(4),
  z1 = rep(3, 4),
  z2 = rep(-2, 4),
  a = c(1, 1, 0, 0),
  b = c(0, 1, 1, 1)
)

# reshape multiple time-varying variables
reshape_longer(
  d,
  columns = list(c("y1", "y2"), c("z1", "z2")),
  names_to = "time",
  values_to = c("new_y", "new_z")
)

}
